requirements.txt
fastapi
uvicorn[standard]
langchain
openai
python-dotenv
streamlit
requests